feat: add guard globally
