--- src/auth/constants/index.ts
@@ -1,1 +1,0 @@
-export { jwtConstants } from './constants';
