--- src/auth/login/dto/create-login.dto.ts
@@ -1,11 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class LoginDto {
-  @IsNotEmpty()
-  @IsString()
-  login: string;
-
-  @IsNotEmpty()
-  @IsString()
-  password: string;
-}
