--- src/auth/refresh/dto/create-refresh.dto.ts
@@ -1,1 +1,0 @@
-export class CreateRefreshDto {}
