--- src/auth/refresh/dto/update-refresh.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateRefreshDto } from './create-refresh.dto';
-
-export class UpdateRefreshDto extends PartialType(CreateRefreshDto) {}
