--- src/auth/refresh/entities/refresh.entity.ts
@@ -1,1 +1,0 @@
-export class Refresh {}
