--- src/auth/refresh/refresh.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { RefreshService } from './refresh.service';
-import { RefreshController } from './refresh.controller';
-
-@Module({
-  controllers: [RefreshController],
-  providers: [RefreshService],
-})
-export class RefreshModule {}
