--- src/auth/signup/entities/signup.entity.ts
@@ -1,1 +1,0 @@
-export class Signup {}
